+++ src/main/java/com/untill/driver/interfaces/bills/VoucherBillClosingRequest.java
@@ -1,48 +1,19 @@
 package com.untill.driver.interfaces.bills;
 
-import com.untill.driver.interfaces.cash.ICashHandler;
-
-import java.math.BigDecimal;
+import java.util.List;
 
 /**
  * A request for {@link ICustomBillsHandler} <tt>purchase voucher</tt> operation
  */
 public class VoucherBillClosingRequest extends CustomBillsHandlerRequest {
 
-	/**
-	 * Voucher identifier
-	 */
-	private String voucherId;
-	/**
-	 * Voucher amount
-	 */
-	private BigDecimal amount;
-	/**
-	 * Payment kind
-	 */
-	private Integer paymentKind;
+	private List<Voucher> vouchers;
 
-	public String getVoucherId() {
-		return voucherId;
+	public List<Voucher> getVouchers() {
+		return vouchers;
 	}
 
-	public void setVoucherId(String voucherId) {
-		this.voucherId = voucherId;
-	}
-
-	public BigDecimal getAmount() {
-		return amount;
-	}
-
-	public void setAmount(BigDecimal amount) {
-		this.amount = amount;
-	}
-
-	public Integer getPaymentKind() {
-		return paymentKind;
-	}
-
-	public void setPaymentKind(Integer paymentKind) {
-		this.paymentKind = paymentKind;
+	public void setVouchers(List<Voucher> vouchers) {
+		this.vouchers = vouchers;
 	}
 }
